update more emoji
