+++ components/pages/testimonials/TestimonialPersonInfo.tsx
@@ -5,7 +5,7 @@ import ImageComponent from '@/components/shared/ImageComponent'
 import ExternalLink from '@/components/shared/ExternalLink'
 import { ImageLoading } from '@/lib/utils/typeDefinitions/enums'
 import { TestimonialPersonInfoProps } from '@/lib/utils/typeDefinitions/props/pages/testimonials'
-import { TEXT } from '@/localization/english'
+import { ARIA_LABELS, TEXT } from '@/localization/english'
 
 const IMAGE_WIDTH = 48
 const IMAGE_HEIGHT = 48
@@ -31,7 +31,7 @@ const TestimonialPersonInfo: FC<TestimonialPersonInfoProps> = ({
     <div className="flex flex-col items-start">
       <div className="flex flex-col font-medium text-gray-900 sm:flex-row sm:items-center">
         <span>
-          <span role="img" className="select-none">
+          <span role="img" aria-label={ARIA_LABELS.quote} className="select-none">
             {icon}
           </span>{' '}
           {name}
