+++ lib/utils/typeDefinitions/props/shared/breadcrumbs-item.ts
@@ -1,0 +1,8 @@
+// Breadcrumbs Item Props
+export type BreadcrumbsItemProps = {
+  href: string
+  icon: string
+  ariaLabel: string
+  text: string
+  showSeparatorIcon?: boolean
+}
